Allele auf double geändert

+++ src/main/main/Evolution.cs
@@ -22,8 +22,8 @@ namespace main
 		public int countChilds;							// Anzahl zu erzeugender Kinder
 		public double recombinationProbability;			// Rekombinationswahrscheinlichkeit
 		public bool InvertOnMutate;						// Mutation: true = invertieren, false = tauschen
-		public int minAllelValue;						// Minimaler Wert der Allele
-		public int maxAllelValue;						// Maximaler Wert der Allele
+		public double minAllelValue;						// Minimaler Wert der Allele
+		public double maxAllelValue;						// Maximaler Wert der Allele
 		public Helper.Enums.SelPropType SelPropType;	// Selektionswahrscheinlichkeit: Fitness- oder Rangbasiert
 		public Helper.Enums.SelType SelType;			// Selektionsverfahren: Roulette, Single-, MultiTournament
 		public Helper.Enums.Encryption Encryption;		// Genomkodierung
@@ -88,47 +88,7 @@ namespace main
 				}
 			}			
 		}
-		
-		/// <summary>
-		/// Mutiert ein Genom
-		/// </summary>
-		/// <param name="genome">Das Genome, was mutiert werden soll</param>
-//		private void Mutate (List<Genome> genomes) 
-//		{
-//			foreach (Genome genome in genomes)
-//			{				
-//				int z1 = 0;
-//				int z2 = 0;
-//				int tmp = 0;
-//				int length = genome.Count;
-//				bool equal = true;
-//	
-//				//Zwei verschiedene Zufallsindices ermitteln
-//				while (equal)
-//				{
-//					//Von Index 1 an da sich der erste Wert (index 0) nicht ändern soll
-//					z1 = Helper.GetRandomInteger(1, length-1);
-//					z2 = Helper.GetRandomInteger(1, length-1);
-//					if (z1 != z2 )
-//						equal = false;
-//				}
-//	
-//				// Wenn true, invertiere, sonst, tausche
-//				if (InvertOnMutate) 
-//				{
-//					tmp = z1 < z2 ? z1 : z2;
-//					genome.Reverse(tmp, Math.Abs(z1-z2)+1);
-//				}
-//				else
-//				{
-//					tmp = genome[z1];
-//					genome[z1] = genome[z2];
-//					genome[z2] = tmp;
-//				}
-//			}
-//			return;
-//		}
-		
+
 		/// <summary>
 		/// Der eigentliche evolutionäre Algorithmus - entspricht doc/EvoAlgTSP.pdf.
 		/// </summary>
